Python/Django/Jai_Ecommerce/requirements.txt
Django
djangorestframework
snowflake-connector-python
pandas
matplotlib
plotly
